Bug 1660470 - Split SurfaceDescriptor.h from ShadowLayerUtils.h. r=mattwoodrow
Differential Revision: https://phabricator.services.mozilla.com/D93561
