+++ gfx/layers/moz.build
@@ -83,6 +83,7 @@ if CONFIG["MOZ_WIDGET_TOOLKIT"] == "windows":
 
 EXPORTS.gfxipc += [
     "ipc/ShadowLayerUtils.h",
+    "ipc/SurfaceDescriptor.h",
 ]
 
 EXPORTS.mozilla.dom += [
